chore: rename feedsidebar component
